Atualizações 2018 02

+++ app/src/main/java/br/edu/utfpr/comunicacaowebmobile/model/servidor/PedidoItem.kt
@@ -1,6 +1,4 @@
 package br.edu.utfpr.comunicacaowebmobile.model.servidor
 
-import java.math.BigDecimal
-
 data class PedidoItem(var codigo: Int, var pedido: Pedido, var produto: Produto,
-                      var quantidade: BigDecimal, var valorUnitario: BigDecimal)
+                      var quantidade: Double, var valorUnitario: Double)
